backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database - Neon Postgres
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Vector Database - Qdrant
qdrant-client==1.7.3

# Embeddings - Cohere API (FREE tier: 100 req/min)
# Get free key at: https://dashboard.cohere.com/api-keys
# Uses httpx (already included below)

# LLM - Groq API (FREE tier)
groq==0.4.1

# Tokenization (for chunking)
tiktoken==0.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client
httpx==0.26.0

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5

# Code quality
black==24.2.0
flake8==7.0.0
mypy==1.8.0

# CORS
fastapi-cors==0.0.6

# Rate limiting
slowapi==0.1.9

# Logging
python-json-logger==2.0.7
